requirements.txt
pandas
matplotlib
requests
prophet
scikit-learn
openpyxl
python-dotenv
